refactored ownership code into network folder
